requirements.txt
# Azure Machine Learning SDK v2 dependencies
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
azure-storage-blob>=12.19.0

# ML libraries
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Web framework for serving
flask>=2.3.0
gunicorn>=21.0.0

# Utilities
pyyaml>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0